last commit

+++ lib/src/model/track_one.dart
@@ -15,18 +15,18 @@ class Model {
 
   Model(
       {this.coord,
-        this.weather,
-        this.base,
-        this.main,
-        this.visibility,
-        this.wind,
-        this.clouds,
-        this.dt,
-        this.sys,
-        this.timezone,
-        this.id,
-        this.name,
-        this.cod});
+      this.weather,
+      this.base,
+      this.main,
+      this.visibility,
+      this.wind,
+      this.clouds,
+      this.dt,
+      this.sys,
+      this.timezone,
+      this.id,
+      this.name,
+      this.cod});
 
   Model.fromJson(Map<String, dynamic> json) {
     coord = json['coord'] != null ? new Coord.fromJson(json['coord']) : null;
@@ -41,7 +41,7 @@ class Model {
     visibility = json['visibility'];
     wind = json['wind'] != null ? new Wind.fromJson(json['wind']) : null;
     clouds =
-    json['clouds'] != null ? new Clouds.fromJson(json['clouds']) : null;
+        json['clouds'] != null ? new Clouds.fromJson(json['clouds']) : null;
     dt = json['dt'];
     sys = json['sys'] != null ? new Sys.fromJson(json['sys']) : null;
     timezone = json['timezone'];
@@ -135,11 +135,11 @@ class Main {
 
   Main(
       {this.temp,
-        this.feelsLike,
-        this.tempMin,
-        this.tempMax,
-        this.pressure,
-        this.humidity});
+      this.feelsLike,
+      this.tempMin,
+      this.tempMax,
+      this.pressure,
+      this.humidity});
 
   Main.fromJson(Map<String, dynamic> json) {
     temp = json['temp'];
@@ -163,7 +163,7 @@ class Main {
 }
 
 class Wind {
-  int? speed;
+  double? speed;
   int? deg;
 
   Wind({this.speed, this.deg});
